# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAmSurfer API",
    "version": "1",
    "description": "API REST versionada (/api/v1) do app mobile IAmSurfer. Autenticação via JWT Bearer (access 15min + refresh 30d rotativo). Erros no formato { \"error\": { \"code\", \"message\", \"fields\" } }. Feeds paginados por cursor: { \"items\": [...], \"next_cursor\": ... }."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/webhooks/payments": {
      "post": {
        "operationId": "post_api_photos_api_payments_webhook",
        "tags": [
          "photos"
        ],
        "summary": "Confirma pagamentos. Público (não usa JWT) e IDEMPOTENTE.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/purchases": {
      "get": {
        "operationId": "get_api_photos_api_my_purchases",
        "tags": [
          "photos"
        ],
        "summary": "GET /me/purchases",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/photo-sessions": {
      "get": {
        "operationId": "get_api_photos_api_list_sessions",
        "tags": [
          "photos"
        ],
        "summary": "GET /photo-sessions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/ranking": {
      "get": {
        "operationId": "get_api_ranking_api_ranking",
        "tags": [
          "ranking"
        ],
        "summary": "Ranking por XP (offset/limit). Cada item traz `rank` e `points`.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/photos/{photo_id}/purchase": {
      "post": {
        "operationId": "post_api_photos_api_purchase_photo",
        "tags": [
          "photos"
        ],
        "summary": "POST /photos/{photo_id}/purchase",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/users/{username}/badges": {
      "get": {
        "operationId": "get_api_ranking_api_user_badges",
        "tags": [
          "ranking"
        ],
        "summary": "Selos de papel ativos do usuário (fotógrafo, atleta, etc.).",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/photo-sessions/{session_id}": {
      "get": {
        "operationId": "get_api_photos_api_get_session",
        "tags": [
          "photos"
        ],
        "summary": "GET /photo-sessions/{session_id}",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Formato de erro único da API.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código estável, legível por máquina."
              },
              "message": {
                "type": "string",
                "description": "Mensagem humana (pt-BR)."
              },
              "fields": {
                "type": "object",
                "description": "Erros de validação por campo.",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}